# core-lending-supply-chain-finance-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-supply-chain-finance API
  description: Supply Chain Finance Management System (Factoring & Confirming)
  contact:
    name: Firefly Software Foundation
    url: https://getfirefly.io
    email: dev@getfirefly.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
...
paths:
...
  /api/v1/supply-chain-finance/fees/{id}:
    get:
      tags:
        - SupplyChainFinanceFee
      summary: Get a supply chain finance fee by ID
      description: Retrieve a specific fee by its unique identifier
      operationId: getById_2
      parameters:
        - name: id
          in: path
          description: Unique identifier of the fee
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Fee found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceFeeDTO'
        '404':
          description: Fee not found
    put:
      tags:
        - SupplyChainFinanceFee
      summary: Update an existing supply chain finance fee
      description: Update the details of an existing fee
      operationId: update_2
      parameters:
        - name: id
          in: path
          description: Unique identifier of the fee
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceFeeDTO'
        required: true
      responses:
        '200':
          description: Fee updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceFeeDTO'
        '400':
          description: Invalid fee data
        '404':
          description: Fee not found
    delete:
      tags:
        - SupplyChainFinanceFee
      summary: Delete a supply chain finance fee
      description: Delete a fee from the system
      operationId: delete_2
      parameters:
        - name: id
          in: path
          description: Unique identifier of the fee
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Fee deleted successfully
        '404':
          description: Fee not found
...
  /api/v1/supply-chain-finance/agreements/{id}:
    get:
      tags:
        - SupplyChainFinanceAgreement
      summary: Get a supply chain finance agreement by ID
      description: Retrieve a specific agreement by its unique identifier
      operationId: getById_4
      parameters:
        - name: id
          in: path
          description: Unique identifier of the agreement
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Agreement found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceAgreementDTO'
        '404':
          description: Agreement not found
    put:
      tags:
        - SupplyChainFinanceAgreement
      summary: Update an existing supply chain finance agreement
      description: Update the details of an existing agreement
      operationId: update_4
      parameters:
        - name: id
          in: path
          description: Unique identifier of the agreement
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceAgreementDTO'
        required: true
      responses:
        '200':
          description: Agreement updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceAgreementDTO'
        '400':
          description: Invalid agreement data
        '404':
          description: Agreement not found
    delete:
      tags:
        - SupplyChainFinanceAgreement
      summary: Delete a supply chain finance agreement
      description: Delete an agreement from the system
      operationId: delete_4
      parameters:
        - name: id
          in: path
          description: Unique identifier of the agreement
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Agreement deleted successfully
        '404':
          description: Agreement not found
...
  /api/v1/supply-chain-finance/fees:
    get:
      tags:
        - SupplyChainFinanceFee
      summary: List/Search supply chain finance fees
      description: Retrieve a paginated list of fees based on filter criteria
      operationId: findAll_2
      parameters:
        - name: filterRequest
          in: query
          description: Filter criteria for fees
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestSupplyChainFinanceFeeDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved fees
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
    post:
      tags:
        - SupplyChainFinanceFee
      summary: Create a new supply chain finance fee
      description: Create a new fee in the system
      operationId: create_2
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceFeeDTO'
        required: true
      responses:
        '200':
          description: Fee created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceFeeDTO'
        '400':
          description: Invalid fee data
...
  /api/v1/supply-chain-finance/agreements:
    get:
      tags:
        - SupplyChainFinanceAgreement
      summary: List/Search supply chain finance agreements
      description: Retrieve a paginated list of supply chain finance agreements based on filter criteria
      operationId: findAll_4
      parameters:
        - name: filterRequest
          in: query
          description: Filter criteria for agreements
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestSupplyChainFinanceAgreementDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved agreements
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
    post:
      tags:
        - SupplyChainFinanceAgreement
      summary: Create a new supply chain finance agreement
      description: Create a new supply chain finance agreement in the system
      operationId: create_4
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceAgreementDTO'
        required: true
      responses:
        '200':
          description: Agreement created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceAgreementDTO'
        '400':
          description: Invalid agreement data
...
components:
  schemas:
...
    SupplyChainFinanceFeeDTO:
      required:
        - agreementId
        - currency
        - feeType
        - financeType
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        financeType:
          type: string
          enum:
            - FACTORING
            - CONFIRMING
        agreementId:
          type: string
          format: uuid
        feeType:
          type: string
          enum:
            - DISCOUNT_FEE
            - SERVICE_FEE
            - ADMINISTRATION_FEE
            - VERIFICATION_FEE
            - EARLY_PAYMENT_FEE
            - COLLECTION_FEE
            - LATE_PAYMENT_FEE
            - OTHER
        currency:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
            - JPY
            - CNY
        fixedAmount:
          type: number
        percentageRate:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        minimumAmount:
          type: number
        maximumAmount:
          type: number
        description:
          maxLength: 500
          minLength: 0
          type: string
...
    SupplyChainFinanceAgreementDTO:
      required:
        - agreementNumber
        - currency
        - financeType
        - loanServicingCaseId
        - startDate
        - status
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        financeType:
          type: string
          enum:
            - FACTORING
            - CONFIRMING
        agreementNumber:
          maxLength: 100
          minLength: 0
          type: string
        loanServicingCaseId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - DRAFT
            - PENDING_APPROVAL
            - ACTIVE
            - SUSPENDED
            - TERMINATED
            - CLOSED
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        currency:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
            - JPY
            - CNY
        creditLimit:
          type: number
        availableLimit:
          type: number
        outstandingAmount:
          type: number
        recourse:
          type: boolean
        notificationRequired:
          type: boolean
        advanceRate:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        supplierEarlyPaymentOption:
          type: boolean
        standardPaymentTermDays:
          type: integer
          format: int32
        earlyPaymentDiscountDays:
          type: integer
          format: int32
        interestRate:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        notes:
          maxLength: 2000
          minLength: 0
          type: string
        createdBy:
          type: string
          format: uuid
        lastModifiedBy:
          type: string
          format: uuid
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestSupplyChainFinanceFeeDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SupplyChainFinanceFeeDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    FilterRequestSupplyChainFinanceAgreementDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SupplyChainFinanceAgreementDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...